ss-py/pyproject.toml
[project]
name = "ss-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ebooklib>=0.20",
    "pillow>=12.0.0",
    "result>=0.17.0",
]
